requirements.txt
numpy
h5py
matplotlib
pyyaml
Pillow
opencv-python
tensorflow
flask
